Move pieces into model.pieces
